modify Android build configuration so that the C# class hierarchy more closely resembles the Java class hierarchy.
This is done by setting _EnableInterfaceMembers to true in Android binding project.

https://docs.microsoft.com/en-us/xamarin/android/release-notes/10/10.1#c-80-default-interface-members-for-bindings-projects
